# 🔒 Enterprise PostgreSQL Security Hardening & Container UID Standardization