--- a/lightrag/config/__init__.py
+++ b/lightrag/config/__init__.py
@@ -1,0 +1,11 @@
+"""
+Configuration management for LightRAG batching orchestrator.
+"""
+
+from .batch_config import BatchingConfig, get_production_config, validate_environment
+
+__all__ = [
+    "BatchingConfig",
+    "get_production_config", 
+    "validate_environment"
+]