--- a/lightrag/orchestrator/__init__.py
+++ b/lightrag/orchestrator/__init__.py
@@ -1,0 +1,20 @@
+"""
+Ollama Batching Orchestrator for LightRAG.
+
+Provides intelligent batching of embedding requests to improve performance
+and resource utilization when using Ollama embeddings.
+"""
+
+from .batch_orchestrator import OllamaBatchingOrchestrator
+from .request_queue import RequestQueue
+from .embedding_cache import EmbeddingCache
+from .batch_processor import BatchProcessor
+from .error_handler import BatchProcessingErrorHandler
+
+__all__ = [
+    "OllamaBatchingOrchestrator",
+    "RequestQueue", 
+    "EmbeddingCache",
+    "BatchProcessor",
+    "BatchProcessingErrorHandler"
+]