… .github/workflows/security-scan.yml
… docker-compose.enhanced.yml
… docker-compose.production.yml
… docs/CHROMADB_ARCHITECTURE_EVALUATION.md
… docs/OLLAMA_BATCHING_ORCHESTRATOR.md
… docs/POSTGRESQL_IMAGE_SECURITY_ADVISORY.md
… docs/POSTGRESQL_SECURITY_FIXES_IMPLEMENTATION.md
… docs/POSTGRESQL_SECURITY_REVIEW.md
… docs/POSTGRESQL_VS_CHROMADB_COMPARISON.md
… docs/UID_MIGRATION_SUMMARY.md
… docs/security/SECURITY_HARDENING.md
… examples/ollama_batching_demo.py
… lightrag/lightrag.py
… lightrag/operate.py
… lightrag/orchestrator/batch_orchestrator.py
… lightrag/orchestrator/batch_processor.py
… lightrag/orchestrator/embedding_cache.py
… lightrag/orchestrator/error_handler.py
… lightrag/orchestrator/request_queue.py
… lightrag/pipeline/batch_integration.py
… lightrag/pipeline/enhanced_embedding.py
… lightrag/resilience.py
… lightrag/xai_optimizer.py
… lightrag_mcp/client/direct_client.py
… postgres/config/postgresql-ssl.conf
… postgres/init/00-remove-default-users.sh
… postgres/init/02-password-security.sh
… postgres/init/03-ssl-setup.sh
… postgres/ssl/generate-ssl-certs.sh
… production.env